+++ android/app/src/main/kotlin/deckers/thibault/aves/channel/calls/StorageHandler.kt
@@ -135,7 +135,7 @@ class StorageHandler(private val context: Context) : MethodCallHandler {
         }
 
         val trashDirs = context.getExternalFilesDirs(null).mapNotNull { StorageUtils.trashDirFor(context, it.path) }
-        val trashItemPaths = trashDirs.flatMap { dir -> dir.listFiles()?.mapNotNull { file -> file?.path } ?: listOf() }
+        val trashItemPaths = trashDirs.flatMap { dir -> dir.listFiles()?.filterNotNull()?.mapNotNull { file -> file.path } ?: listOf() }
         val untrackedPaths = trashItemPaths.filterNot(knownPaths::contains).toList()
 
         result.success(untrackedPaths)