+++ android/app/src/main/kotlin/deckers/thibault/aves/channel/streams/ActivityResultStreamHandler.kt
@@ -198,7 +198,7 @@ class ActivityResultStreamHandler(private val activity: Activity, arguments: Any
             .setDataAndType(AppAdapterHandler.getShareableUri(activity, Uri.parse(uri)), mimeType)
 
         if (intent.resolveActivity(activity.packageManager) == null) {
-            error("edit-resolve", "cannot resolve activity for this intent", null)
+            error("edit-resolve", "cannot resolve activity for this intent for uri=$uri mimeType=$mimeType", null)
             return
         }
 
@@ -207,7 +207,7 @@ class ActivityResultStreamHandler(private val activity: Activity, arguments: Any
             endOfStream()
         }
         if (!safeStartActivityForResult(intent, MainActivity.EDIT_REQUEST)) {
-            error("edit-start", "cannot start activity for this intent", null)
+            error("edit-start", "cannot start activity for this intent for uri=$uri mimeType=$mimeType", null)
         }
     }