+++ plugins/aves_video/lib/aves_video.dart
@@ -1,5 +1,6 @@
 library aves_video;
 
+export 'src/ab_repeat.dart';
 export 'src/controller.dart';
 export 'src/metadata.dart';
 export 'src/settings/subtitles.dart';